+++ src/sas/qtgui/Perspectives/Corfunc/CorfuncPerspective.py
@@ -5,20 +5,21 @@ This module provides the intelligence behind the gui interface for Corfunc.
 
 # global
 from PyQt5.QtGui import QStandardItem
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg \
-    as FigureCanvas
 from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT
-from matplotlib.figure import Figure
+
 from numpy.linalg.linalg import LinAlgError
 import numpy as np
 
-from typing import Optional, List, Tuple
+from typing import Optional, List
 
 from PyQt5 import QtCore
 from PyQt5 import QtGui, QtWidgets
 
 # sas-global
 # pylint: disable=import-error, no-name-in-module
+
+from sas.qtgui.Perspectives.Corfunc.CorfuncSlider import CorfuncSlider
+from sas.qtgui.Perspectives.Corfunc.CorfuncCanvas import CorfuncCanvas
 
 import sas.qtgui.Utilities.GuiUtils as GuiUtils
 from sas.qtgui.Utilities.Reports.reportdata import ReportData
@@ -32,167 +33,6 @@ from .UI.CorfuncPanel import Ui_CorfuncDialog
 from .corefuncutil import WIDGETS
 from .saveextrapolated import SaveExtrapolatedPopup
 from ..perspective import Perspective
-
-class MyMplCanvas(FigureCanvas):
-    """Ultimately, this is a QWidget (as well as a FigureCanvasAgg, etc.)."""
-
-    def __init__(self, model, width=5, height=4, dpi=100):
-        self.model = model
-        self.fig = Figure(figsize=(width, height), dpi=dpi)
-        self.axes = self.fig.add_subplot(111)
-
-        FigureCanvas.__init__(self, self.fig)
-
-        self.data: Optional[Tuple[np.ndarray, np.ndarray, np.ndarray]] = None
-        self.extrap = None
-        self.dragging = None
-        self.draggable = False
-        self.leftdown = False
-        self.fig.canvas.mpl_connect("button_release_event", self.on_mouse_up)
-        self.fig.canvas.mpl_connect("button_press_event", self.on_mouse_down)
-        self.fig.canvas.mpl_connect('motion_notify_event', self.on_motion)
-
-    def on_legend(self, qx, qy):
-        """
-        Checks if mouse coursor is on legend box
-        :return:
-        """
-        on_legend_box = False
-        bbox = self.legend.get_window_extent()
-        if qx > bbox.xmin and qx < bbox.xmax and qy > bbox.ymin  and qy < bbox.ymax:
-            on_legend_box = True
-        return  on_legend_box
-
-    def on_mouse_down(self, event):
-        if not self.draggable:
-            return
-        if event.button == 1:
-            self.leftdown = True
-        if self.on_legend(event.x, event.y):
-            return
-
-        qmin = float(self.model.item(WIDGETS.W_QMIN).text())
-        qmax1 = float(self.model.item(WIDGETS.W_QMAX).text())
-        qmax2 = float(self.model.item(WIDGETS.W_QCUTOFF).text())
-
-        q = event.xdata
-
-        if (np.abs(q-qmin) < np.abs(q-qmax1) and
-            np.abs(q-qmin) < np.abs(q-qmax2)):
-            self.dragging = "qmin"
-        elif (np.abs(q-qmax2) < np.abs(q-qmax1)):
-            self.dragging = "qmax2"
-        else:
-            self.dragging = "qmax1"
-
-    def on_mouse_up(self, event):
-        if not self.dragging:
-            return None
-        if event.button == 1:
-            self.leftdown = False
-        if self.on_legend(event.x, event.y):
-            return
-
-        if self.dragging == "qmin":
-            item = WIDGETS.W_QMIN
-        elif self.dragging == "qmax1":
-            item = WIDGETS.W_QMAX
-        else:
-            item = WIDGETS.W_QCUTOFF
-
-        self.model.setItem(item, QtGui.QStandardItem(str(GuiUtils.formatNumber(event.xdata))))
-
-        self.dragging = None
-
-    def on_motion(self, event):
-        if not self.leftdown:
-            return
-        if not self.draggable:
-            return
-        if self.dragging is None:
-            return
-
-        if self.dragging == "qmin":
-            item = WIDGETS.W_QMIN
-        elif self.dragging == "qmax1":
-            item = WIDGETS.W_QMAX
-        else:
-            item = WIDGETS.W_QCUTOFF
-
-        self.model.setItem(item, QtGui.QStandardItem(str(GuiUtils.formatNumber(event.xdata))))
-
-
-    def draw_q_space(self):
-        """Draw the Q space data in the plot window
-
-        This draws the q space data in self.data, as well
-        as the bounds set by self.qmin, self.qmax1, and self.qmax2.
-        It will also plot the extrpolation in self.extrap, if it exists."""
-
-        self.draggable = True
-
-        self.fig.clf()
-
-        self.axes = self.fig.add_subplot(111)
-        self.axes.set_xscale("log")
-        self.axes.set_yscale("log")
-        self.axes.set_xlabel("Q [$\AA^{-1}$]")
-        self.axes.set_ylabel("I(Q) [cm$^{-1}$]")
-        self.axes.set_title("Scattering data")
-        self.fig.tight_layout()
-
-        qmin = float(self.model.item(WIDGETS.W_QMIN).text())
-        qmax1 = float(self.model.item(WIDGETS.W_QMAX).text())
-        qmax2 = float(self.model.item(WIDGETS.W_QCUTOFF).text())
-
-        if self.data:
-            # self.axes.plot(self.data.x, self.data.y, label="Experimental Data")
-            self.axes.errorbar(self.data.x, self.data.y, yerr=self.data.dy, label="Experimental Data")
-            self.axes.axvline(qmin)
-            self.axes.axvline(qmax1)
-            self.axes.axvline(qmax2)
-            self.axes.set_xlim(min(self.data.x) / 2,
-                               max(self.data.x) * 1.5 - 0.5 * min(self.data.x))
-            self.axes.set_ylim(min(self.data.y) / 2,
-                               max(self.data.y) * 1.5 - 0.5 * min(self.data.y))
-
-        if self.extrap:
-            self.axes.plot(self.extrap.x, self.extrap.y, label="Extrapolation")
-
-        if self.data or self.extrap:
-            self.legend = self.axes.legend()
-
-        self.draw()
-
-    def draw_real_space(self):
-        """
-        This function draws the real space data onto the plot
-
-        The 1d correlation function in self.data, the 3d correlation function
-        in self.data3, and the interface distribution function in self.data_idf
-        are all draw in on the plot in linear cooredinates."""
-
-        self.draggable = False
-
-        self.fig.clf()
-
-        self.axes = self.fig.add_subplot(111)
-        self.axes.set_xscale("linear")
-        self.axes.set_yscale("linear")
-        self.axes.set_xlabel("Z [$\AA$]")
-        self.axes.set_ylabel("Correlation")
-        self.axes.set_title("Real Space Correlations")
-        self.fig.tight_layout()
-
-        if self.data:
-            data1, data3, data_idf = self.data
-            self.axes.plot(data1.x, data1.y, label="1D Correlation")
-            self.axes.plot(data3.x, data3.y, label="3D Correlation")
-            self.axes.set_xlim(0, max(data1.x) / 4)
-            self.legend = self.axes.legend()
-
-        self.draw()
-
 
 class CorfuncWindow(QtWidgets.QDialog, Ui_CorfuncDialog, Perspective):
     """Displays the correlation function analysis of sas data."""
@@ -229,12 +69,17 @@ class CorfuncWindow(QtWidgets.QDialog, Ui_CorfuncDialog, Perspective):
         self.txtLowerQMin.setEnabled(False)
         self.extrapolation_curve = None
 
-        self._q_space_plot = MyMplCanvas(self.model)
-        self.plotLayout.insertWidget(0, self._q_space_plot)
-        self.plotLayout.insertWidget(1, NavigationToolbar2QT(self._q_space_plot, self))
-        self._real_space_plot = MyMplCanvas(self.model) # TODO: This is not a good name, or structure either
-        self.plotLayout.insertWidget(2, self._real_space_plot)
-        self.plotLayout.insertWidget(3, NavigationToolbar2QT(self._real_space_plot, self))
+
+        self.slider = CorfuncSlider()
+        self.plotLayout.insertWidget(2, self.slider)
+
+        self._q_space_plot = CorfuncCanvas(self.model)
+        self.plotLayout.insertWidget(1, self._q_space_plot)
+        self.plotLayout.insertWidget(2, NavigationToolbar2QT(self._q_space_plot, self))
+
+        self._real_space_plot = CorfuncCanvas(self.model)
+        self.plotLayout.insertWidget(3, self._real_space_plot)
+        self.plotLayout.insertWidget(4, NavigationToolbar2QT(self._real_space_plot, self))
 
         self.gridLayout_4.setColumnStretch(0, 1)
         self.gridLayout_4.setColumnStretch(1, 2)
@@ -319,6 +164,7 @@ class CorfuncWindow(QtWidgets.QDialog, Ui_CorfuncDialog, Perspective):
         self._real_space_plot.data = None
         self._real_space_plot.extrap = None
         self._real_space_plot.draw_real_space()
+        self.slider.setEnabled(False)
         # Clear calculator, model, and data path
         self._calculator = CorfuncCalculator()
         self._model_item = None
@@ -333,6 +179,7 @@ class CorfuncWindow(QtWidgets.QDialog, Ui_CorfuncDialog, Perspective):
             return
         self.mapper.toFirst()
         self._q_space_plot.draw_q_space()
+        #TODO: Update slider
 
     def _update_calculator(self):
         self._calculator.lowerq = float(self.model.item(WIDGETS.W_QMIN).text())
